+++ protocol/rest/server/server_impl/go_restful_server_test.go
@@ -18,6 +18,7 @@
 package server_impl
 
 import (
+	"dubbo.apache.org/dubbo-go/v3/config"
 	"testing"
 )
 
@@ -27,7 +28,6 @@ import (
 
 import (
 	"dubbo.apache.org/dubbo-go/v3/common"
-	"dubbo.apache.org/dubbo-go/v3/protocol/rest/config"
 	"dubbo.apache.org/dubbo-go/v3/protocol/rest/server"
 )
